app/emails.py
from flask.ext.mail import Message
from app import mail
from decorators import async

#@async
def send_async_email(msg):
   mail.send(msg)


def send_email(subject, sender, recipients, text_body, html_body):
    msg = Message(subject = subject, sender = sender, recipients = recipients)
    msg.body = text_body
    msg.html = html_body
    send_async_email(msg)