requirements.txt
Babel==1.3
Flask==0.10
Flask-Babel==0.9
Flask-Login==0.2.7
Flask-Mail==0.9.0
Flask-OpenID==1.1.1
Flask-SQLAlchemy==0.16
Flask-WTF==0.9.3
Flask-WhooshAlchemy==0.54a
Jinja2==2.7.1
MarkupSafe==0.18
SQLAlchemy==0.7.9
Tempita==0.5.1
WTForms==1.0.5
Werkzeug==0.9.4
Whoosh==2.5.4
argparse==1.2.1
blinker==1.3
decorator==3.4.0
flup==1.0.2
gunicorn==18.0
itsdangerous==0.23
python-openid==2.2.5
pytz==2013b
speaklater==1.3
sqlalchemy-migrate==0.7.2
wsgiref==0.1.2
psycopg2==2.5
gunicorn==0.17.2